fix rather stupid bug causing images to reset improperly and other images to not size properly
Apparently 2011 me thought it was possible to index {} using objects as
keys instead of strings. This was causing all imageData[] lookups to
return the same data, no matter what. So, when images were reset to
their original size, they always inherited the size of the first image
that was resized. And when trying to interact with other images on the
page before resetting all images, all images thought they were currently
resized.

Also refactored a bunch of code: more functions, less duplication.

Thanks again to Reddit user /u/huck_ for the bug reports. Not sure how
these slipped through the cracks for so long!